+++ src/public/Core/Global.h
@@ -2,10 +2,10 @@
 
 class Global {
 public:
-	static const double FIXED_DELTA_TIME;
+	static const float FIXED_DELTA_TIME;
 	static const  int SCREEN_WIDTH = 1024;
 	static const int SCREEN_HEIGHT = 576;
-	static const double SCREEN_FORMAT;
-	static const  double METER;
+	static const float SCREEN_FORMAT;
+	static const  float METER;
 	static const bool DEBUG = true;
 };
